Add support for manually added students
This feature *requires* the Canvas code to be patched to expose the sis_batch_id
in enrollments. If this patch is not made, this script would now treat all existing
users as manually enrolled and re-enroll them every time it's run

Manually added students have no sis_batch_id. Such students are stored in a hash and
after processing a course, any manually added students who are now in the SIS feed for
any section in the course are deleted from their section *if* the section they were
in is different from the one the SIS places them in. If the section was the same, we
assume Canvas will convert them to SIS enrollments (i.e give them an sis_batch_id)