Rewrite Manual Student processing
Manually added students can't be deleted via the sis_import method,
so collect mnaually added students as an array of enrollment objects
per course. At the end of each course, see if any manually added students
match the SIS source for any of the sections in the course and if they do
iterate through the array of enrollments and do a DELETE enrollment API call
for each. Confused yet?