Send pushAccountUpdate amaint DA for each newly created user
This commit modifies the user creation process minimly in order to ensure that newly-created Canvas users have communication channels set up for their SFU aliases. Rather than completely refactor the process to remove the SIS API call and fire CREATE_CANVAS_USER messages to be consumed by amaint-canvas-jms, this does the minimum amount necessary.

The process-enrollments script waits for the SIS import to complete and then re-fetches all Canvas users for its internal cache. This means that firing a CREATE_CANVAS_USER for each user-to-be-added won't work well; that process is completely fire-and-forget; there's nothing for process-enrollments to check to see if the creates succeeded before re-fetching all users.

So, instead, we keep the SIS import, and then after the import completes successfully, fire an pushAccountUpdate direct action through amaint to trigger a syncLogin message for each new user. This message will be consumed by amaint-canvas-jms, which will see that the user exists in Canvas, and see that it needs Communication Channels created for its aliases, and will do so.